[FP-22] Rename: CardDataList.ts 파일 이름을 reservationCardsData.ts로 변경
