+++ src/components/Cards/reservationApprovalCard.tsx
@@ -1,0 +1,5 @@
+const ReservationApprovalOrRejectionCard = () => {
+  return <div>hi</div>;
+};
+
+export default ReservationApprovalOrRejectionCard;
